requirements_extract.txt
prettyprint
python-dotenv==1.0.1
tqdm==4.66.2
pandas==2.2.3
numpy
loguru
requests==2.27.1
SQLAlchemy==2.0.25
Flask
Flask-SQLAlchemy
psycopg2-binary

confluent-kafka
redis

opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
#opentelemetry-instrumentation
#opentelemetry-exporter-otlp
#opentelemetry-api 
opentelemetry-sdk  
opentelemetry-instrumentation-system-metrics